proyecto/requirements.txt

django-ckeditor==6.2.0
mysqlclient==2.2.0
sqlparse==0.4.4
asgiref==3.7.2
Pillow==10.0.1
tzdata==2023.3
pylint-django
pylint-celery
whitenoise>=6.0
pylint